+++ Test/output_test.py
@@ -76,10 +76,10 @@ True
 >>> save_obj=cm.save_obj("test",address=False)
 >>> save_obj=={'Status': True, 'Message': None}
 True
->>> save_obj=cm.save_obj("test_metrics",address=False,save_metrics=True)
+>>> save_obj=cm.save_obj("test_stat",address=False,save_stat=True)
 >>> save_obj=={'Status': True, 'Message': None}
 True
->>> save_obj=cm.save_obj("test_no_vectors",address=False,save_vectors=False)
+>>> save_obj=cm.save_obj("test_no_vectors",address=False,save_vector=False)
 >>> save_obj=={'Status': True, 'Message': None}
 True
 >>> cm_file=ConfusionMatrix(file=open("test.obj","r"))
@@ -216,9 +216,9 @@ Y(Youden index)                                                  None           
 dInd(Distance index)                                             None                    0.67315                 0.66926                 1.0
 sInd(Similarity index)                                           None                    0.52401                 0.52676                 0.29289
 <BLANKLINE>
->>> cm_metrics_file=ConfusionMatrix(file=open("test_metrics.obj","r"))
+>>> cm_stat_file=ConfusionMatrix(file=open("test_stat.obj","r"))
 >>> cm_no_vectors_file=ConfusionMatrix(file=open("test_no_vectors.obj","r"))
->>> cm_metrics_file==cm_file
+>>> cm_stat_file==cm_file
 True
 >>> cm_no_vectors_file==cm_file
 True
@@ -227,10 +227,10 @@ True
 True
 >>> cm_no_vectors_dict["Predict-Vector"] == None
 True
->>> cm_metrics_dict = json.load(open("test_metrics.obj","r"))
->>> cm_metrics_dict["Class-Stat"]["MCC"] != None
+>>> cm_stat_dict = json.load(open("test_stat.obj","r"))
+>>> cm_stat_dict["Class-Stat"]["MCC"] != None
 True
->>> cm_metrics_dict["Overall-Stat"]["Overall MCC"] != None
+>>> cm_stat_dict["Overall-Stat"]["Overall MCC"] != None
 True
 >>> def activation(i):
 ...	    if i<0.7:
@@ -422,7 +422,7 @@ True
 >>> os.remove("test_normalized.csv")
 >>> os.remove("test_normalized_matrix.csv")
 >>> os.remove("test.obj")
->>> os.remove("test_metrics.obj")
+>>> os.remove("test_stat.obj")
 >>> os.remove("test_no_vectors.obj")
 >>> os.remove("test.html")
 >>> os.remove("test_normalized.html")